+++ packages/utils-react/src/use-debounce.tsx
@@ -1,0 +1,22 @@
+import { debounce } from "@hiogawa/utils";
+import React from "react";
+import { useStableCallback } from "./utils";
+
+export function useDebounce<F extends (...args: any[]) => void>(
+  f: F,
+  ms: number
+): [F, { isPending: boolean }] {
+  f = useStableCallback(f);
+
+  const [isPending, setIsPending] = React.useState(false);
+
+  const debounced = React.useCallback(
+    debounce(f, ms, {
+      onStart: () => setIsPending(true),
+      onFinish: () => setIsPending(false),
+    }),
+    [ms]
+  );
+
+  return [debounced, { isPending }];
+}
