+++ packages/utils-react/src/utils.ts
@@ -1,4 +1,3 @@
-import { debounce } from "@hiogawa/utils";
 import React from "react";
 
 export function usePrevious<T>(value: T): T {
@@ -53,23 +52,4 @@ export function useRefCallbackEffect<T>(
   };
 
   return React.useCallback(refCallback, []);
-}
-
-export function useDebounce<F extends (...args: any[]) => void>(
-  f: F,
-  ms: number
-): [F, { isPending: boolean }] {
-  f = useStableCallback(f);
-
-  const [isPending, setIsPending] = React.useState(false);
-
-  const debounced = React.useCallback(
-    debounce(f, ms, {
-      onStart: () => setIsPending(true),
-      onFinish: () => setIsPending(false),
-    }),
-    [ms]
-  );
-
-  return [debounced, { isPending }];
 }
